# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "BFD",
    "contact": {}
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/usertype/get": {
      "post": {
        "summary": "Get User Types",
        "tags": [
          "User Type"
        ],
        "operationId": "GetUserTypes",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetUserTypesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usertype": {
      "get": {
        "summary": "Get User Type",
        "tags": [
          "User Type"
        ],
        "operationId": "GetUserType",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/usertype/update": {
      "post": {
        "summary": "Update User Type",
        "tags": [
          "User Type"
        ],
        "operationId": "UpdateUserType",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateUserTypeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/user/get": {
      "post": {
        "summary": "Get Users",
        "tags": [
          "User"
        ],
        "operationId": "GetUsers",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/GetUsersRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user": {
      "get": {
        "summary": "Get User",
        "tags": [
          "User"
        ],
        "operationId": "GetUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/user/update": {
      "post": {
        "summary": "Update User",
        "tags": [
          "User"
        ],
        "operationId": "UpdateUser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "firstName",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "lastName",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "phone",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "address[streetNo]",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "address[city]",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "address[zip]",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "profilePicture",
            "in": "formData",
            "required": true,
            "type": "file",
            "format": "file",
            "description": ""
          },
          {
            "name": "id",
            "in": "formData",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/user/onboarding": {
      "get": {
        "summary": "Onboarding",
        "tags": [
          "User"
        ],
        "operationId": "Onboarding",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "GetUserTypesRequest": {
      "title": "GetUserTypesRequest",
      "example": {
        "sortBy": "",
        "order": "",
        "search": "",
        "page": 1,
        "limit": 10,
        "organizer": ""
      },
      "type": "object",
      "properties": {
        "sortBy": {
          "type": "string"
        },
        "order": {
          "type": "string"
        },
        "search": {
          "type": "string"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "organizer": {
          "type": "string"
        }
      },
      "required": [
        "sortBy",
        "order",
        "search",
        "page",
        "limit",
        "organizer"
      ]
    },
    "UpdateUserTypeRequest": {
      "title": "UpdateUserTypeRequest",
      "example": {
        "id": "",
        "name": "",
        "organizer": "",
        "menu": [
          {
            "menu": "",
            "permissions": {
              "hasAccess": true,
              "canCreate": true,
              "canUpdate": true,
              "canDelete": true
            }
          },
          {
            "menu": "",
            "permissions": {
              "hasAccess": true,
              "canCreate": true,
              "canUpdate": true,
              "canDelete": true
            },
            "children": [
              {
                "menu": "",
                "permissions": {
                  "hasAccess": true,
                  "canCreate": true,
                  "canUpdate": true,
                  "canDelete": true
                }
              }
            ]
          }
        ]
      },
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "organizer": {
          "type": "string"
        },
        "menu": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Menu"
          }
        }
      },
      "required": [
        "id",
        "name",
        "organizer",
        "menu"
      ]
    },
    "Menu": {
      "title": "Menu",
      "example": {
        "menu": "",
        "permissions": {
          "hasAccess": true,
          "canCreate": true,
          "canUpdate": true,
          "canDelete": true
        }
      },
      "type": "object",
      "properties": {
        "menu": {
          "type": "string"
        },
        "permissions": {
          "$ref": "#/definitions/Permissions"
        },
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Child1"
          }
        }
      },
      "required": [
        "menu",
        "permissions"
      ]
    },
    "Permissions": {
      "title": "Permissions",
      "example": {
        "hasAccess": true,
        "canCreate": true,
        "canUpdate": true,
        "canDelete": true
      },
      "type": "object",
      "properties": {
        "hasAccess": {
          "type": "boolean"
        },
        "canCreate": {
          "type": "boolean"
        },
        "canUpdate": {
          "type": "boolean"
        },
        "canDelete": {
          "type": "boolean"
        }
      },
      "required": [
        "hasAccess",
        "canCreate",
        "canUpdate",
        "canDelete"
      ]
    },
    "Child1": {
      "title": "Child1",
      "example": {
        "menu": "",
        "permissions": {
          "hasAccess": true,
          "canCreate": true,
          "canUpdate": true,
          "canDelete": true
        }
      },
      "type": "object",
      "properties": {
        "menu": {
          "type": "string"
        },
        "permissions": {
          "$ref": "#/definitions/Permissions"
        }
      },
      "required": [
        "menu",
        "permissions"
      ]
    },
...
    "GetUsersRequest": {
      "title": "GetUsersRequest",
      "example": {
        "sortBy": "",
        "order": "",
        "search": "",
        "page": 1,
        "limit": 10
      },
      "type": "object",
      "properties": {
        "sortBy": {
          "type": "string"
        },
        "order": {
          "type": "string"
        },
        "search": {
          "type": "string"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "sortBy",
        "order",
        "search",
        "page",
        "limit"
      ]
    },
...
  },
...
}